refactor(android): merge most app list logic
