+++ composeApp/src/androidMain/kotlin/fr/husi/ui/AppListViewModel.kt
@@ -1,29 +1,13 @@
 package fr.husi.ui
 
-import android.content.pm.PackageInfo
 import android.content.pm.PackageManager
-import android.graphics.drawable.Drawable
-import androidx.collection.ArraySet
-import androidx.compose.foundation.text.input.TextFieldState
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.Stable
-import androidx.compose.runtime.snapshotFlow
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import fr.husi.ktx.blankAsNull
-import fr.husi.repository.androidRepo
-import fr.husi.utils.PackageCache
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.currentCoroutineContext
-import kotlinx.coroutines.ensureActive
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
-import kotlinx.coroutines.flow.distinctUntilChanged
-import kotlinx.coroutines.flow.drop
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import fr.husi.resources.*
 
 @Immutable
 internal data class AppListUiState(
@@ -33,37 +17,18 @@ internal data class AppListUiState(
 )
 
 @Stable
-internal class AppListViewModel : ViewModel() {
+internal class AppListViewModel : BaseAppListViewModel() {
     private val _uiState = MutableStateFlow(AppListUiState())
     val uiState = _uiState.asStateFlow()
 
-    val textFieldState = TextFieldState()
-
-    private lateinit var packageManager: PackageManager
-
-    private val proxiedUids = ArraySet<Int>()
-    private val cachedApps by lazy {
-        PackageCache.installedPackages.toMutableMap().apply {
-            remove(androidRepo.context.packageName)
-        }
-    }
-
     init {
-        viewModelScope.launch {
-            snapshotFlow { textFieldState.text.toString() }
-                .drop(1)
-                .distinctUntilChanged()
-                .collect { reload() }
-        }
+        collectSearchText()
     }
 
     fun initialize(pm: PackageManager, packages: Set<String>) {
         packageManager = pm
-
-        viewModelScope.launch(Dispatchers.IO) {
-            _uiState.update {
-                it.copy(isLoading = true, apps = emptyList())
-            }
+        viewModelScope.launch(singleThreadContext) {
+            _uiState.update { it.copy(isLoading = true, apps = emptyList()) }
             proxiedUids.clear()
             val cachedApps = cachedApps
             for ((packageName, packageInfo) in cachedApps) {
@@ -75,149 +40,36 @@ internal class AppListViewModel : ViewModel() {
         }
     }
 
-    private val iconCache = mutableMapOf<String, Drawable>()
-    private fun loadIcon(packageManager: PackageManager, packageInfo: PackageInfo): Drawable {
-        return iconCache.getOrPut(packageInfo.packageName) {
-            packageInfo.applicationInfo!!.loadIcon(packageManager)
-        }
+    override fun updateApps(apps: List<ProxiedApp>, isLoading: Boolean) {
+        _uiState.update { it.copy(isLoading = isLoading, apps = apps) }
     }
 
-    suspend fun reload(cachedApps: Map<String, PackageInfo> = this.cachedApps) {
-        val apps = mutableListOf<ProxiedApp>()
-        val query = textFieldState.text.toString()
-        for ((packageName, packageInfo) in cachedApps) {
-            currentCoroutineContext()[Job]!!.ensureActive()
+    override fun updateSnackbar(message: StringOrRes?) {
+        _uiState.update { it.copy(snackbarMessage = message) }
+    }
 
-            val applicationInfo = packageInfo.applicationInfo!!
-            val name = applicationInfo.loadLabel(packageManager).toString()
-            query.blankAsNull()?.lowercase()?.let {
-                val hit = packageName.lowercase().contains(it)
-                        || name.lowercase().contains(it)
-                        || applicationInfo.uid.toString().contains(it)
-                if (!hit) continue
-            }
+    override suspend fun afterMutation() = reload()
 
-            val proxiedApp = ProxiedApp(
-                appInfo = applicationInfo,
-                packageName = packageName,
-                isProxied = proxiedUids.contains(applicationInfo.uid),
-                icon = loadIcon(packageManager, packageInfo),
-                name = name,
-            )
-            apps.add(proxiedApp)
-        }
-        val comparator = compareBy<ProxiedApp>({ !it.isProxied }, { it.name })
-        apps.sortWith(comparator)
-        _uiState.update {
-            it.copy(
-                isLoading = false,
-                apps = apps,
+    override suspend fun afterItemClick(app: ProxiedApp, newIsProxied: Boolean) {
+        _uiState.update { state ->
+            state.copy(
+                apps = state.apps.map {
+                    if (it.uid == app.uid) it.copy(isProxied = newIsProxied) else it
+                },
             )
         }
-    }
-
-    fun invertSections() = viewModelScope.launch {
-        val current = ArraySet(proxiedUids) // clone
-        val cachedApps = cachedApps
-        val allUids = ArraySet<Int>()
-        for ((_, packageInfo) in cachedApps) {
-            allUids.add(packageInfo.applicationInfo!!.uid)
-        }
-        proxiedUids.clear()
-        proxiedUids.ensureCapacity(cachedApps.size - current.size)
-        allUids.filter { it !in current }.forEach { proxiedUids.add(it) }
-        reload()
-    }
-
-    fun clearSections() = viewModelScope.launch {
-        proxiedUids.clear()
-        reload()
     }
 
     fun allPackages(): ArrayList<String> {
         return cachedApps.mapNotNullTo(ArrayList()) { (packageName, packageInfo) ->
             val uid = packageInfo.applicationInfo!!.uid
-            if (uid in proxiedUids) {
-                packageName
-            } else {
-                null
-            }
+            if (uid in proxiedUids) packageName else null
         }
     }
 
-    fun export(): String {
+    override fun export(): String {
         val body = allPackages().joinToString("\n")
-        val full = "false\n$body"
-        return full
-    }
-
-    fun import(raw: String?) = viewModelScope.launch {
-        if (raw?.blankAsNull() == null) {
-            _uiState.update { state ->
-                state.copy(
-                    snackbarMessage = StringOrRes.Res(Res.string.action_import_err),
-                )
-            }
-            return@launch
-        }
-        var bypass = false
-        val apps = raw.lineSequence().let {
-            when (it.firstOrNull()) {
-                "false" -> {
-                    bypass = false
-                    it.drop(1)
-                }
-
-                "true" -> {
-                    bypass = true
-                    it.drop(1)
-                }
-
-                else -> it
-            }
-        }
-        proxiedUids.clear()
-        val cachedApps = cachedApps
-        val proxiedPackageName = mutableListOf<String>()
-        // Bypass mode: choose all apps that not mentioned in list
-        if (bypass) {
-            for ((packageName, packageInfo) in cachedApps) {
-                if (packageName !in apps) {
-                    proxiedUids.add(packageInfo.applicationInfo!!.uid)
-                    proxiedPackageName.add(packageName)
-                }
-            }
-        } else {
-            for (packageName in apps) {
-                val info = cachedApps[packageName] ?: continue
-                proxiedUids.add(info.applicationInfo!!.uid)
-                proxiedPackageName.add(packageName)
-            }
-        }
-        _uiState.update { state ->
-            state.copy(
-                snackbarMessage = StringOrRes.Res(Res.string.action_import_msg),
-            )
-        }
-        reload()
-    }
-
-    fun onItemClick(app: ProxiedApp) = viewModelScope.launch {
-        val newIsProxied = !proxiedUids.remove(app.uid)
-        if (newIsProxied) {
-            proxiedUids.add(app.uid)
-        }
-        _uiState.update { state ->
-            state.copy(
-                apps = state.apps.map {
-                    if (it.uid == app.uid) {
-                        it.copy(isProxied = newIsProxied)
-                    } else {
-                        it
-                    }
-                },
-            )
-        }
+        return "false\n$body"
     }
 
 }
